+++ examples/basic/simple-post.php
@@ -3,7 +3,7 @@
 require __DIR__ . '/../../vendor/autoload.php';
 
 use PFinal\AsyncioHttp\Client;
-use function PFinal\Asyncio\run;
+use function PfinalClub\Asyncio\run;
 
 function main(): void
 {